// gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Connection = {
  /** A list of edges (relational context) between connected nodes */
  edges: Array<Edge>;
  /** A list of connected nodes */
  nodes: Array<Node>;
  /** Information about pagination in a connection. */
  pageInfo: PageInfo;
};
// ...
export type DatabaseIdentifier = {
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int']['output'];
};
// ...
export type Edge = {
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor: Maybe<Scalars['String']['output']>;
  /** The connected node */
  node: Node;
};
// ...
export type Group = DatabaseIdentifier & MenuItemLinkable & Node & TermNode & UniformResourceIdentifiable & {
  __typename?: 'Group';
  /** Connection between the Group type and the ContentNode type */
  contentNodes: Maybe<GroupToContentNodeConnection>;
  /** The number of objects connected to the object */
  count: Maybe<Scalars['Int']['output']>;
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int']['output'];
  /** The description of the object */
  description: Maybe<Scalars['String']['output']>;
  /** Connection between the TermNode type and the EnqueuedScript type */
  enqueuedScripts: Maybe<TermNodeToEnqueuedScriptConnection>;
  /** Connection between the TermNode type and the EnqueuedStylesheet type */
  enqueuedStylesheets: Maybe<TermNodeToEnqueuedStylesheetConnection>;
  /**
   * The id field matches the WP_Post-&gt;ID field.
   * @deprecated Deprecated in favor of databaseId
   */
  groupId: Maybe<Scalars['Int']['output']>;
  /** The unique resource identifier path */
  id: Scalars['ID']['output'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean']['output'];
  /** Whether the object is restricted from the current viewer */
  isRestricted: Maybe<Scalars['Boolean']['output']>;
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean']['output'];
  /** The link to the term */
  link: Maybe<Scalars['String']['output']>;
  /** The human friendly name of the object. */
  name: Maybe<Scalars['String']['output']>;
  /** Connection between the Group type and the Project type */
  projects: Maybe<GroupToProjectConnection>;
  /** An alphanumeric identifier for the object unique to its type. */
  slug: Maybe<Scalars['String']['output']>;
  /** Connection between the Group type and the Taxonomy type */
  taxonomy: Maybe<GroupToTaxonomyConnectionEdge>;
  /** The name of the taxonomy that the object is associated with */
  taxonomyName: Maybe<Scalars['String']['output']>;
  /** The ID of the term group that this term object belongs to */
  termGroupId: Maybe<Scalars['Int']['output']>;
  /** The taxonomy ID that the object is associated with */
  termTaxonomyId: Maybe<Scalars['Int']['output']>;
  /** The unique resource identifier path */
  uri: Maybe<Scalars['String']['output']>;
};
// ...
export type GroupConnection = {
  /** A list of edges (relational context) between RootQuery and connected Group Nodes */
  edges: Array<GroupConnectionEdge>;
  /** A list of connected Group Nodes */
  nodes: Array<Group>;
  /** Information about pagination in a connection. */
  pageInfo: GroupConnectionPageInfo;
};
// ...
export type NodeWithFeaturedImageToMediaItemConnectionEdge = Edge & MediaItemConnectionEdge & OneToOneConnection & {
  __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor: Maybe<Scalars['String']['output']>;
  /** The node of the connection, without the edges */
  node: MediaItem;
};
// ...
export type Project = ContentNode & DatabaseIdentifier & MenuItemLinkable & Node & NodeWithFeaturedImage & NodeWithTemplate & NodeWithTitle & Previewable & UniformResourceIdentifiable & {
  __typename?: 'Project';
  /** Connection between the ContentNode type and the ContentType type */
  contentType: Maybe<ContentNodeToContentTypeConnectionEdge>;
  /** The name of the Content Type the node belongs to */
  contentTypeName: Scalars['String']['output'];
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int']['output'];
  /** Post publishing date. */
  date: Maybe<Scalars['String']['output']>;
  /** The publishing date set in GMT. */
  dateGmt: Maybe<Scalars['String']['output']>;
  /** The desired slug of the post */
  desiredSlug: Maybe<Scalars['String']['output']>;
  /** If a user has edited the node within the past 15 seconds, this will return the user that last edited. Null if the edit lock doesn&#039;t exist or is greater than 15 seconds */
  editingLockedBy: Maybe<ContentNodeToEditLockConnectionEdge>;
  /** The RSS enclosure for the object */
  enclosure: Maybe<Scalars['String']['output']>;
  /** Connection between the ContentNode type and the EnqueuedScript type */
  enqueuedScripts: Maybe<ContentNodeToEnqueuedScriptConnection>;
  /** Connection between the ContentNode type and the EnqueuedStylesheet type */
  enqueuedStylesheets: Maybe<ContentNodeToEnqueuedStylesheetConnection>;
  /** Connection between the NodeWithFeaturedImage type and the MediaItem type */
  featuredImage: Maybe<NodeWithFeaturedImageToMediaItemConnectionEdge>;
  /** The database identifier for the featured image node assigned to the content node */
  featuredImageDatabaseId: Maybe<Scalars['Int']['output']>;
  /** Globally unique ID of the featured image assigned to the node */
  featuredImageId: Maybe<Scalars['ID']['output']>;
  /** Connection between the Project type and the Group type */
  groups: Maybe<ProjectToGroupConnection>;
  /** The global unique identifier for this post. This currently matches the value stored in WP_Post-&gt;guid and the guid column in the &quot;post_objects&quot; database table. */
  guid: Maybe<Scalars['String']['output']>;
  /** The globally unique identifier of the projects object. */
  id: Scalars['ID']['output'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean']['output'];
  /** Whether the object is a node in the preview state */
  isPreview: Maybe<Scalars['Boolean']['output']>;
  /** Whether the object is restricted from the current viewer */
  isRestricted: Maybe<Scalars['Boolean']['output']>;
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean']['output'];
  /** The user that most recently edited the node */
  lastEditedBy: Maybe<ContentNodeToEditLastConnectionEdge>;
  /** The permalink of the post */
  link: Maybe<Scalars['String']['output']>;
  /** The local modified time for a post. If a post was recently updated the modified field will change to match the corresponding time. */
  modified: Maybe<Scalars['String']['output']>;
  /** The GMT modified time for a post. If a post was recently updated the modified field will change to match the corresponding time in GMT. */
  modifiedGmt: Maybe<Scalars['String']['output']>;
  /** Connection between the Project type and the Project type */
  preview: Maybe<ProjectToPreviewConnectionEdge>;
  /** The database id of the preview node */
  previewRevisionDatabaseId: Maybe<Scalars['Int']['output']>;
  /** Whether the object is a node in the preview state */
  previewRevisionId: Maybe<Scalars['ID']['output']>;
  /** Added to the GraphQL Schema because the ACF Field Group &quot;Project Fields&quot; was set to Show in GraphQL. */
  projectFields: Maybe<Project_Projectfields>;
  /**
   * The id field matches the WP_Post-&gt;ID field.
   * @deprecated Deprecated in favor of the databaseId field
   */
  projectId: Scalars['Int']['output'];
  /** The uri slug for the post. This is equivalent to the WP_Post-&gt;post_name field and the post_name column in the database for the &quot;post_objects&quot; table. */
  slug: Maybe<Scalars['String']['output']>;
  /** The current status of the object */
  status: Maybe<Scalars['String']['output']>;
  /** The template assigned to the node */
  template: Maybe<ContentTemplate>;
  /** Connection between the Project type and the TermNode type */
  terms: Maybe<ProjectToTermNodeConnection>;
  /** The title of the post. This is currently just the raw title. An amendment to support rendered title needs to be made. */
  title: Maybe<Scalars['String']['output']>;
  /** The unique resource identifier path */
  uri: Maybe<Scalars['String']['output']>;
};
// ...
export type ProjectToGroupConnection = Connection & GroupConnection & {
  __typename?: 'ProjectToGroupConnection';
  /** Edges for the ProjectToGroupConnection connection */
  edges: Array<ProjectToGroupConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Group>;
  /** Information about pagination in a connection. */
  pageInfo: ProjectToGroupConnectionPageInfo;
};
// ...
export type Project_Projectfields = AcfFieldGroup & {
  __typename?: 'Project_Projectfields';
  client: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName: Maybe<Scalars['String']['output']>;
  rating: Maybe<Scalars['Float']['output']>;
};
// ...
export type RootQueryToGroupConnection = Connection & GroupConnection & {
  __typename?: 'RootQueryToGroupConnection';
  /** Edges for the RootQueryToGroupConnection connection */
  edges: Array<RootQueryToGroupConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Group>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToGroupConnectionPageInfo;
};

/** An edge in a connection */
export type RootQueryToGroupConnectionEdge = Edge & GroupConnectionEdge & {
  __typename?: 'RootQueryToGroupConnectionEdge';
  /** A cursor for use in pagination */
  cursor: Maybe<Scalars['String']['output']>;
  /** The item at the end of the edge */
  node: Group;
};

/** Page Info on the &quot;RootQueryToGroupConnection&quot; */
export type RootQueryToGroupConnectionPageInfo = GroupConnectionPageInfo & PageInfo & WpPageInfo & {
  __typename?: 'RootQueryToGroupConnectionPageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor: Maybe<Scalars['String']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor: Maybe<Scalars['String']['output']>;
};
// ...
export type RootQueryToProjectConnection = Connection & ProjectConnection & {
  __typename?: 'RootQueryToProjectConnection';
  /** Edges for the RootQueryToProjectConnection connection */
  edges: Array<RootQueryToProjectConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Project>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToProjectConnectionPageInfo;
};

/** An edge in a connection */
export type RootQueryToProjectConnectionEdge = Edge & ProjectConnectionEdge & {
  __typename?: 'RootQueryToProjectConnectionEdge';
  /** A cursor for use in pagination */
  cursor: Maybe<Scalars['String']['output']>;
  /** The item at the end of the edge */
  node: Project;
};
// ...
export type GroupListQuery = { __typename?: 'RootQuery', groups: { __typename?: 'RootQueryToGroupConnection', edges: Array<{ __typename?: 'RootQueryToGroupConnectionEdge', node: { __typename?: 'Group', databaseId: number, name: string | null, slug: string | null } }> } | null };

export type ProjectListQueryVariables = Exact<{ [key: string]: never; }>;


export type ProjectListQuery = { __typename?: 'RootQuery', projects: { __typename?: 'RootQueryToProjectConnection', edges: Array<{ __typename?: 'RootQueryToProjectConnectionEdge', node: { __typename?: 'Project', databaseId: number, slug: string | null, title: string | null, date: string | null, groups: { __typename?: 'ProjectToGroupConnection', nodes: Array<{ __typename?: 'Group', name: string | null, slug: string | null }> } | null, projectFields: { __typename?: 'Project_Projectfields', rating: number | null, client: string | null } | null, featuredImage: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: { __typename?: 'MediaItem', sourceUrl: string | null } } | null } }> } | null };